Make code more modular (#13)
* Modularizes metrics into separate submodules

* Reformats grade output for consistency

* Refines HTML rendering and updates badge configuration

* Remove colon from branch label in HTML output

* Refactors HTML label assertion in tests

* Enhances color labeling in HTML dashboard

* Adds color and icons to dashboard labels

* Improves HTML rendering by separating icon and label

* Improves label color handling for dashboard

* Refines PR label validation logic

* Enhances PR label checks with push trigger

* Removes push trigger from label workflow
