+++ .badgery.yaml
@@ -10,24 +10,24 @@ cards:
 
   - group: Tests
     type: gh_action
-    title: Test package on PyPI
+    title: Test package (PyPI.org)
     file: test-pypi.yaml
     enabled: true
 
   - group: Code Quality
     type: codefactor
-    title: Code quality from CodeFactor.io
+    title: Code quality (CodeFactor.io)
     enabled: true
 
   - group: Code Quality
     type: radon_mi
-    title: Maintainability index with radon
+    title: Maintainability index (radon)
     report: reports/{branch}/maintainability-index.json
     enabled: true
 
   - group: Code Quality
     type: radon_cc
-    title: Cyclomatic complexity with radon
+    title: Cyclomatic complexity (radon)
     report: reports/{branch}/cyclomatic-complexity.json
     enabled: true
 
@@ -51,12 +51,12 @@ cards:
 
   - group: Coverage
     type: codecov
-    title: Unit test coverage from Codecov.io
+    title: Unit test coverage (Codecov.io)
     enabled: true
 
   - group: Coverage
     type: interrogate
-    title: Docstring coverage with interrogate
+    title: Docstring coverage (interrogate)
     report: reports/{branch}/coverage-docstring.txt
     enabled: true
 
@@ -68,6 +68,6 @@ cards:
 
   - group: Build & Release
     type: gh_action
-    title: Publish to PyPI
+    title: Publish (PyPI.org)
     workflow: publish-pypi.yaml
     enabled: true
